Fix generic typing of Tables.wrap
The T variable for the map values is bound when the method is called.
So the T will always refer exactly to the type of values the compiler
knows the passed object to contain. I think there is no need to use
"? extends T" and I am guessing it could cause problems downstream.