Making guessParser less error-prone
When auto detecting the column type of such a column, the previous
guessParser implementation produced exceptions:

-1
0
-1.0
1.0
0.0
infinity
-infinity
+Nan
NaN
-NaN
1.23e7
13.8f
57269d

Therefore, now the only number format which will be guessed by
guessParser is Double. Also, when testing this with imagej-server, the
Strings "infinity" and "Nan" where used which are not compatible with
Double::valueOf, therefore they are replaced with the compatible
capitalization. Maybe there is a better way to  do this..

Co-authored-by: Jan Eglinger <jan.eglinger@fmi.ch>