Guard against the addition of null columns
And add a quick test that the problem is fixed.

This fix prevents cryptic exceptions like this one:

java.lang.NullPointerException: Cannot invoke "org.scijava.table.Column.size()" because "column" is null
    at org.scijava.table.AbstractTable.add(AbstractTable.java:167)
    at org.scijava.table.AbstractTable.add(AbstractTable.java:44)
    at org.scijava.util.SizableArrayList.setSize(SizableArrayList.java:90)
    at org.scijava.table.AbstractTable.setColumnCount(AbstractTable.java:75)
    at org.scijava.table.AbstractTable.<init>(AbstractTable.java:66)
    at org.scijava.table.DefaultFloatTable.<init>(DefaultFloatTable.java:49)

The above exception happens if too many initial columns are given,
because the SizableArrayList under the hood sometimes tries (if a certain
hack fails) to expand the size of the ArrayList by repeatedly adding nulls.

The new test checks that such tables can be created with lots of
initial columns, *and* that we can explicitly add null columns,
without this sort of exception getting thrown.